requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-frontmatter==1.0.0
markdown-it-py==3.0.0
GitPython==3.1.43
chromadb==0.5.3
chroma-hnswlib==0.7.3
sentence-transformers==2.7.0
rank-bm25==0.2.2
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
SQLAlchemy==2.0.31
httpx==0.27.0
typer==0.12.3
pyyaml==6.0.1
pytest==8.2.2
black==24.4.2
ruff==0.5.5
langgraph==0.0.62
